+++ lib/Logger.php
@@ -1,26 +1,28 @@
 <?php
 namespace yii\mustache;
 
-use yii\base\{InvalidParamException};
+use Psr\Log\{LoggerInterface, LoggerTrait, LogLevel};
+use yii\base\{InvalidParamException, Object};
 use yii\log\{Logger as YiiLogger};
 
 /**
  * Component used to log messages from the view engine to the application logger.
  */
-class Logger extends \Mustache_Logger_AbstractLogger {
+class Logger extends Object implements LoggerInterface {
+  use LoggerTrait;
 
   /**
    * @var int[] Mappings between Mustache levels and Yii ones.
    */
   private static $levels = [
-    \Mustache_Logger::ALERT => YiiLogger::LEVEL_ERROR,
-    \Mustache_Logger::CRITICAL => YiiLogger::LEVEL_ERROR,
-    \Mustache_Logger::DEBUG => YiiLogger::LEVEL_TRACE,
-    \Mustache_Logger::EMERGENCY => YiiLogger::LEVEL_ERROR,
-    \Mustache_Logger::ERROR => YiiLogger::LEVEL_ERROR,
-    \Mustache_Logger::INFO => YiiLogger::LEVEL_INFO,
-    \Mustache_Logger::NOTICE => YiiLogger::LEVEL_INFO,
-    \Mustache_Logger::WARNING => YiiLogger::LEVEL_WARNING
+    LogLevel::ALERT => YiiLogger::LEVEL_ERROR,
+    LogLevel::CRITICAL => YiiLogger::LEVEL_ERROR,
+    LogLevel::DEBUG => YiiLogger::LEVEL_TRACE,
+    LogLevel::EMERGENCY => YiiLogger::LEVEL_ERROR,
+    LogLevel::ERROR => YiiLogger::LEVEL_ERROR,
+    LogLevel::INFO => YiiLogger::LEVEL_INFO,
+    LogLevel::NOTICE => YiiLogger::LEVEL_INFO,
+    LogLevel::WARNING => YiiLogger::LEVEL_WARNING
   ];
 
   /**
@@ -32,7 +34,7 @@ class Logger extends \Mustache_Logger_AbstractLogger {
    */
   public function log($level, $message, array $context = []) {
     if (!isset(static::$levels[$level])) {
-      $values = implode(', ', (new \ReflectionClass(\Mustache_Logger::class))->getConstants());
+      $values = implode(', ', (new \ReflectionClass(LogLevel::class))->getConstants());
       throw new InvalidParamException("Invalid enumerable value \"$level\". Please make sure it is among ($values).");
     }